+++ cli/azd/pkg/tools/gh.go
@@ -12,6 +12,7 @@ import (
 	"strings"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/executil"
+	"github.com/blang/semver/v4"
 )
 
 type GitHubCli interface {
@@ -39,8 +40,35 @@ var (
 
 type ghCli struct{}
 
-func (cli *ghCli) CheckInstalled(_ context.Context) (bool, error) {
-	return toolInPath("gh")
+func (cli *ghCli) versionInfo() VersionInfo {
+	return VersionInfo{
+		MinimumVersion: semver.Version{
+			Major: 2,
+			Minor: 4,
+			Patch: 0},
+		UpdateCommand: "Visit https://github.com/cli/cli/releases to upgrade",
+	}
+}
+
+func (cli *ghCli) CheckInstalled(ctx context.Context) (bool, error) {
+	found, err := toolInPath("gh")
+	if !found {
+		return false, err
+	}
+	ghRes, err := executeCommand(ctx, "gh", "--version")
+	if err != nil {
+		return false, fmt.Errorf("checking %s version: %w", cli.Name(), err)
+	}
+	ghSemver, err := extractSemver(ghRes)
+	if err != nil {
+		return false, fmt.Errorf("converting to semver version fails: %w", err)
+	}
+	updateDetail := cli.versionInfo()
+	if ghSemver.LT(updateDetail.MinimumVersion) {
+		return false, &ErrSemver{ToolName: cli.Name(), versionInfo: updateDetail}
+	}
+
+	return true, nil
 }
 
 func (cli *ghCli) Name() string {