+++ cli/azd/pkg/tools/swa_test.go
@@ -12,66 +12,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func Test_SwaLogin(t *testing.T) {
-	tempSwaCli := NewSwaCli()
-	swacli := tempSwaCli.(*swaCli)
-
-	ran := false
-
-	t.Run("NoErrors", func(t *testing.T) {
-		swacli.runWithResultFn = func(ctx context.Context, args executil.RunArgs) (executil.RunResult, error) {
-			ran = true
-
-			require.Equal(t, []string{
-				"-y", "@azure/static-web-apps-cli",
-				"login",
-				"--tenant-id", "tenantID",
-				"--subscription-id", "subscriptionID",
-				"--resource-group", "resourceGroupID",
-				"--app-name", "appName",
-			}, args.Args)
-
-			return executil.RunResult{
-				Stdout: "",
-				Stderr: "",
-				// if the returned `error` is nil we don't return an error. The underlying 'exec'
-				// returns an error if the command returns a non-zero exit code so we don't actually
-				// need to check it.
-				ExitCode: 1,
-			}, nil
-		}
-
-		err := swacli.Login(context.Background(), "tenantID", "subscriptionID", "resourceGroupID", "appName")
-		require.NoError(t, err)
-		require.True(t, ran)
-	})
-
-	t.Run("Error", func(t *testing.T) {
-		swacli.runWithResultFn = func(ctx context.Context, args executil.RunArgs) (executil.RunResult, error) {
-			ran = true
-
-			require.Equal(t, []string{
-				"-y", "@azure/static-web-apps-cli",
-				"login",
-				"--tenant-id", "tenantID",
-				"--subscription-id", "subscriptionID",
-				"--resource-group", "resourceGroupID",
-				"--app-name", "appName",
-			}, args.Args)
-
-			return executil.RunResult{
-				Stdout:   "stdout text",
-				Stderr:   "stderr text",
-				ExitCode: 1,
-			}, errors.New("example error message")
-		}
-
-		err := swacli.Login(context.Background(), "tenantID", "subscriptionID", "resourceGroupID", "appName")
-		require.True(t, ran)
-		require.EqualError(t, err, "swa login: exit code: 1, stdout: stdout text, stderr: stderr text: example error message")
-	})
-}
-
 func Test_SwaBuild(t *testing.T) {
 	tempSwaCli := NewSwaCli()
 	swacli := tempSwaCli.(*swaCli)
@@ -82,11 +22,11 @@ func Test_SwaBuild(t *testing.T) {
 		swacli.runWithResultFn = func(ctx context.Context, args executil.RunArgs) (executil.RunResult, error) {
 			ran = true
 
-			require.Equal(t, "./appFolderPath", args.Cwd)
+			require.Equal(t, "./projectPath", args.Cwd)
 			require.Equal(t, []string{
-				"-y", "@azure/static-web-apps-cli",
+				"-y", "@azure/static-web-apps-cli@1.0.0",
 				"build",
-				"--app-location", ".",
+				"--app-location", "service/path",
 				"--output-location", "build",
 			}, args.Args)
 
@@ -100,7 +40,7 @@ func Test_SwaBuild(t *testing.T) {
 			}, nil
 		}
 
-		err := swacli.Build(context.Background(), "./appFolderPath", "build")
+		err := swacli.Build(context.Background(), "./projectPath", "service/path", "build")
 		require.NoError(t, err)
 		require.True(t, ran)
 	})
@@ -109,11 +49,11 @@ func Test_SwaBuild(t *testing.T) {
 		swacli.runWithResultFn = func(ctx context.Context, args executil.RunArgs) (executil.RunResult, error) {
 			ran = true
 
-			require.Equal(t, "./appFolderPath", args.Cwd)
+			require.Equal(t, "./projectPath", args.Cwd)
 			require.Equal(t, []string{
-				"-y", "@azure/static-web-apps-cli",
+				"-y", "@azure/static-web-apps-cli@1.0.0",
 				"build",
-				"--app-location", ".",
+				"--app-location", "service/path",
 				"--output-location", "build",
 			}, args.Args)
 
@@ -124,7 +64,7 @@ func Test_SwaBuild(t *testing.T) {
 			}, errors.New("example error message")
 		}
 
-		err := swacli.Build(context.Background(), "./appFolderPath", "build")
+		err := swacli.Build(context.Background(), "./projectPath", "service/path", "build")
 		require.True(t, ran)
 		require.EqualError(t, err, "swa build: exit code: 1, stdout: stdout text, stderr: stderr text: example error message")
 	})
@@ -140,17 +80,19 @@ func Test_SwaDeploy(t *testing.T) {
 		swacli.runWithResultFn = func(ctx context.Context, args executil.RunArgs) (executil.RunResult, error) {
 			ran = true
 
-			require.Equal(t, "./appFolderPath", args.Cwd)
+			require.Equal(t, "./projectPath", args.Cwd)
 			require.Equal(t, []string{
-				"-y", "@azure/static-web-apps-cli",
+				"-y", "@azure/static-web-apps-cli@1.0.0",
 				"deploy",
 				"--tenant-id", "tenantID",
 				"--subscription-id", "subscriptionID",
 				"--resource-group", "resourceGroupID",
 				"--app-name", "appName",
-				"--app-location", ".",
+				"--app-location", "service/path",
 				"--output-location", "build",
-				"--env", "production",
+				"--env", "default",
+				"--no-use-keychain",
+				"--deployment-token", "deploymentToken",
 			}, args.Args)
 
 			return executil.RunResult{
@@ -163,7 +105,7 @@ func Test_SwaDeploy(t *testing.T) {
 			}, nil
 		}
 
-		_, err := swacli.Deploy(context.Background(), "tenantID", "subscriptionID", "resourceGroupID", "appName", "./appFolderPath", "build", "production")
+		_, err := swacli.Deploy(context.Background(), "./projectPath", "tenantID", "subscriptionID", "resourceGroupID", "appName", "service/path", "build", "default", "deploymentToken")
 		require.NoError(t, err)
 		require.True(t, ran)
 	})
@@ -172,17 +114,19 @@ func Test_SwaDeploy(t *testing.T) {
 		swacli.runWithResultFn = func(ctx context.Context, args executil.RunArgs) (executil.RunResult, error) {
 			ran = true
 
-			require.Equal(t, "./appFolderPath", args.Cwd)
+			require.Equal(t, "./projectPath", args.Cwd)
 			require.Equal(t, []string{
-				"-y", "@azure/static-web-apps-cli",
+				"-y", "@azure/static-web-apps-cli@1.0.0",
 				"deploy",
 				"--tenant-id", "tenantID",
 				"--subscription-id", "subscriptionID",
 				"--resource-group", "resourceGroupID",
 				"--app-name", "appName",
-				"--app-location", ".",
+				"--app-location", "service/path",
 				"--output-location", "build",
-				"--env", "production",
+				"--env", "default",
+				"--no-use-keychain",
+				"--deployment-token", "deploymentToken",
 			}, args.Args)
 
 			return executil.RunResult{
@@ -192,7 +136,7 @@ func Test_SwaDeploy(t *testing.T) {
 			}, errors.New("example error message")
 		}
 
-		_, err := swacli.Deploy(context.Background(), "tenantID", "subscriptionID", "resourceGroupID", "appName", "./appFolderPath", "build", "production")
+		_, err := swacli.Deploy(context.Background(), "./projectPath", "tenantID", "subscriptionID", "resourceGroupID", "appName", "service/path", "build", "default", "deploymentToken")
 		require.True(t, ran)
 		require.EqualError(t, err, "swa deploy: exit code: 1, stdout: stdout text, stderr: stderr text: example error message")
 	})