+++ templates/todo/api/csharp-cosmos-sql/Program.cs
@@ -15,8 +15,7 @@ builder.Services.AddSingleton(_ => new CosmosClient(builder.Configuration["AZURE
     }
 }));
 builder.Services.AddControllers();
-var options = new ApplicationInsightsServiceOptions { ConnectionString = builder.Configuration["APPLICATIONINSIGHTS_CONNECTION_STRING"] };
-builder.Services.AddApplicationInsightsTelemetry(options);
+builder.Services.AddApplicationInsightsTelemetry(builder.Configuration);
 
 var app = builder.Build();